Add files via upload
This folder contains the main.cpp file needed to configure the ESP8266 to the desired wifi network, and show which IP address hosts the server. Change the SSID and Password to the appropriate network. This program must be run the first time connecting to a wifi network. The terminal will print an IP address indicating where the server will be hosted.